backend/requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
httpx>=0.27.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncmy>=0.2.9
redis>=5.0.0
greenlet>=3.0.0
bcrypt>=4.0.1
